# backEnd/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Media API",
        "description": "Documentation for Social Media API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/club-admins": {
            "get": {
                "tags": [
                    "ClubAdmin"
                ],
                "summary": "Get all club admins",
                "description": "Retrieves a list of all club admins",
                "operationId": "99f71b07251c8b6d358a54c58d50af60",
                "responses": {
                    "200": {
                        "description": "List of club admins",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClubAdminResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ClubAdmin"
                ],
                "summary": "Create a new club admin",
                "description": "Creates a new club admin with the provided data",
                "operationId": "cfd15aed83a48a4090b9012718463677",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomClub": {
                                        "description": "The name of the club",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the club",
                                        "type": "string"
                                    },
                                    "ecole": {
                                        "description": "The school associated with the club",
                                        "type": "string"
                                    },
                                    "tactique": {
                                        "description": "The tactical approach of the club",
                                        "type": "string"
                                    },
                                    "gestion": {
                                        "description": "The management style of the club",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Club admin created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClubAdminResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/club-admins/user/{userId}": {
            "get": {
                "tags": [
                    "ClubAdmin"
                ],
                "summary": "Get club admin by user ID",
                "description": "Retrieves a club admin based on the associated user ID",
                "operationId": "f2c190973f59a11725873ada19448fa8",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The ID of the user associated with the club admin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Club admin retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClubAdminResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club admin not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ClubAdmin"
                ],
                "summary": "Delete club admin by user ID",
                "description": "Deletes a club admin based on the associated user ID",
                "operationId": "47aaa180595c72dc2605558312ad2fe2",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The ID of the user associated with the club admin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Club admin deleted successfully"
                    },
                    "404": {
                        "description": "Club admin not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/club-admins/user": {
            "put": {
                "tags": [
                    "ClubAdmin"
                ],
                "summary": "Update club admin by user ID",
                "description": "Updates a club admin based on the associated user ID",
                "operationId": "48f061507a3da21e6ff706191df1815d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomClub": {
                                        "description": "The name of the club",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the club",
                                        "type": "string"
                                    },
                                    "ecole": {
                                        "description": "The school associated with the club",
                                        "type": "string"
                                    },
                                    "tactique": {
                                        "description": "The tactical approach of the club",
                                        "type": "string"
                                    },
                                    "gestion": {
                                        "description": "The management style of the club",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Club admin updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClubAdminResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club admin not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/club-admins/search/{name}": {
            "get": {
                "tags": [
                    "ClubAdmin"
                ],
                "summary": "Search club admins by name",
                "description": "Searches for club admins by their name",
                "operationId": "ea220a8295e4032d4bb1577b3685ed15",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Club admins found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClubAdminResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No club admins found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClubAdminResource": {
                "title": "ClubAdminResource",
                "description": "Resource representing a club admin",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "nomClub": {
                        "type": "string",
                        "example": "FC Barcelona"
                    },
                    "description": {
                        "type": "string",
                        "example": "Club admin for FC Barcelona"
                    },
                    "ecole": {
                        "type": "string",
                        "example": "Ecole de Foot"
                    },
                    "tactique": {
                        "type": "string",
                        "example": "4-4-2"
                    },
                    "gestion": {
                        "type": "string",
                        "example": "Professional management"
                    },
                    "titres": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TitreResource"
                        }
                    },
                    "joueurs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JoueurResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "ExperienceResource": {
                "title": "ExperienceResource",
                "description": "Resource representing a player's experience",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nomClub": {
                        "type": "string",
                        "example": "FC Barcelona"
                    },
                    "joinDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-01"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-12-31"
                    },
                    "place": {
                        "type": "string",
                        "example": "Spain"
                    },
                    "categoryType": {
                        "type": "string",
                        "example": "Professional"
                    },
                    "joueur_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "JoueurResource": {
                "title": "JoueurResource",
                "description": "Resource representing a player",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "experiences": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExperienceResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "TitreResource": {
                "title": "TitreResource",
                "description": "Resource representing a title won by a club admin",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nomTitre": {
                        "type": "string",
                        "example": "UEFA Champions League"
                    },
                    "description": {
                        "type": "string",
                        "example": "Won the UEFA Champions League in 2020"
                    },
                    "number": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "Resource representation of a User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "bio": {
                        "type": "string",
                        "example": "This is John's bio"
                    },
                    "profileImage": {
                        "type": "string",
                        "example": "http://localhost:8081/storage/profile_images/john.jpg"
                    },
                    "bannerImage": {
                        "type": "string",
                        "example": "http://localhost:8081/storage/banner_images/john.jpg"
                    },
                    "role": {
                        "type": "string",
                        "example": "Joueur"
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    },
                    "stats": {
                        "properties": {
                            "followers_count": {
                                "type": "integer",
                                "example": 100
                            },
                            "following_count": {
                                "type": "integer",
                                "example": 50
                            },
                            "posts_count": {
                                "type": "integer",
                                "example": 20
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}